requirements.txt
pandas
pytest
streamlit
plotly
google-genai
python-dotenv
